Added tests, the Project class, and the ProjectSet class
I've opted to whip up an incredibly simple testing class to aid in TDD for my own sanity. I wanted to avoid needing to add bundler to this project since one of my goals is simplicity.
The first pass implementation of the Project class is complete, as of now I wanted to create a class that understands the concept of duration and full / tavel days. It is possible that not all the methods I've defined so far will need to be used in the final implementation, but I wanted to get my thoughts down in case they came in handy later.
I've introduced the ProjectSet class which will be in charge of doing the reimbursement calculation. I will handle the nitty gritty of that in the next PR.
Finally, I've added some xtests (skipped tests) to show the direction I'm headed in for my next effort.