Implemented project set total calculation
I've removed a few methods on Project that I did not end up using, updated tests, and implmented the logic behind calculating a project sets total reimbursement. There are a few assumptions I made that still need to be ironed out, specifically around date neighbors, and a few method names could be made clearer.
Additionally, the main.rb entrypoint needs to be written, but this will look very similar to the last tests in the projet_set_tests file.