Updated README, modified project set start and end date logic to force travel day rates
The README now has a visualization to help understand the logic behind the rate calculations.
I've also added the program output so curious watchers don't need to run the code locally.
The logic surrounding travel days has changed slightly so that the first and last days of a project set are forced to be travel days, regardless of overlaps or neighbors.
Tests have been updated to reflect this change.